nightly-build/docker-compose-nexus.yml: add depends_on for kong, kong-migrations
Note that docker-compose's definition of depends_on is just that the container
was started, i.e. the process was forked and started executing, so this doesn't
solve dependency problems on it's own, but I think it helps in getting a more
efficient startup ordering.

Also kong-migrations and kong should depend on consul so they can access the
scripts from the consul-scripts volume and use consul to query the state of
services for proper dependency waiting.

Signed-off-by: Ian Johnson <ian.johnson@canonical.com>