nightly-build/docker-compose-nexus.yml: use consul script for sec svcs
The consul svc status script from docker-edgex-consul can now be used to
wait for services to be ready as defined by Consul before running their
full command. Note we use entrypoint to override a hard-coded ENTRYPOINT
in the image, we can't use command like we do for the other containers
(i.e. kong and vault and postgres, etc.)

Also add the vault-config volume to consul so that consul can check for
the existence of the TLS files.

Signed-off-by: Ian Johnson <ian.johnson@canonical.com>