Enable edgex-proxy to use per-service token instead of root token.
Modifies the secrets volume of edgex-proxy so that it has access
to its per-service token.  Remove the volume mount to the actual
TLS certificate since edgex-proxy gets the TLS cert from the
secret store.  Fix dependency on edgex-proxy (actually depends
on vault-worker to upload the TLS cert, not vault) and
remove kong-db (kong already depends on kong-db).

Signed-off-by: Bryon Nevis <bryon.nevis@intel.com>