fix: Consul health check runtime config
Added the new environment variables to the consul services in each of
the compose files. The vars, EDGEX_DB and EDGEX_SECURE are utilized in
an entrypoint script in the docker-edgex-consul image that initializes
default config and script files based on these environment variables.
These variables are NOT required (as they have defaults if the system
can't find them), however, it is encouraged to use them so that
configuration is explicitly set, and the developers happen to know what
might be going on in the event of a misconfigured consul service.

Signed-off-by: Beau Frusetta <beau.frusetta@intel.com>