fix: Make Kong datastore persistent and bump Postgres version
The docker implementation of Kong nejects to make its
Postgres database persistent.  Under a variety of scenarios,
this can result in Kong forgetting authentication credentials,
or user-applied customizations.  This condition is also
inconsistent with the snap implementation, which
currently persists Postgres data.

This fix creates a persistent docker volume and maps
it to the PGDATA directory of the Postgres container.

This fix also performs a proactive minor version bump of Postgres.

Signed-off-by: Bryon Nevis <bryon.nevis@intel.com>