feat(security): Deploy new security-bootstrapper service
Docker-compose deploys with a new security-bootstrapper service which controls the security bootstrapping steps for various phases.
The details are summarized in ADR secure bootstrapping.

Add command env addition and command overrides for core-services

Now the environment vars for security-stage-gate are in env file and overrides when necessary.

Update asc-http-export-secure and asc-mqtt-export-secure to be gated by security-bootstrapper.

Added common-sec-stage-gate.env description

Updated some typos in the document

Standardize the security-bootstrapper env file and naming of some docker containers
eg vault-worker -> secretstore-setup, edgex-proxy -> proxy-setup

Use the name common-sec-stage-gate.env as the env file for security-bootstrapper

Fix the problem of missing common.env and env quotation

Standardize the naming for appservice to app-service

Closes: #349, #237

Signed-off-by: Jim Wang <yutsung.jim.wang@intel.com>